+++ app/src/js/chaos/renderers/WebGLRenderer.js
@@ -1,12 +1,63 @@
-let Renderer = require('../Renderer'),
-    Three = require('three'),
-    RenderFilter = require('./webgl/RenderFilter'),
-    PencilSketchFilter = require('./webgl/PencilSketchFilter'),
+let m4 = require('../../utils/matrix'),
+    Renderer = require('../Renderer'),
     Components = require('../Components'),
     Props = require('../Props');
 
-class WebGLRenderer extends Renderer {
-    static get displayName() { return "WebGL Renderer"; }
+const VERTEX_PROGRAM = `
+uniform mat4 rotation;
+uniform mat4 projection;
+attribute vec4 position;
+attribute vec4 color;
+
+varying vec4 vColor;
+
+void main() {
+    gl_Position = projection * rotation * position;
+    gl_PointSize = 1.0;
+    vColor = color;
+}
+`;
+
+const FRAGMENT_PROGRAM = `
+precision mediump float;
+
+varying vec4 vColor;
+
+void main() {
+    gl_FragColor = vColor;
+}
+`;
+
+function degToRad(d) {
+    return d * Math.PI / 180;
+}
+
+function createShader(gl, type, source) {
+    const shader = gl.createShader(type);
+    gl.shaderSource(shader, source);
+    gl.compileShader(shader);
+    return shader;
+}
+
+function createVertexShader(gl) {
+    return createShader(gl, gl.VERTEX_SHADER, VERTEX_PROGRAM);
+}
+
+function createFragmentShader(gl) {
+    return createShader(gl, gl.FRAGMENT_SHADER, FRAGMENT_PROGRAM);
+}
+
+function createShaderProgram(gl) {
+    const program = gl.createProgram();
+    gl.attachShader(program, createVertexShader(gl));
+    gl.attachShader(program, createFragmentShader(gl));
+    gl.linkProgram(program);
+    gl.useProgram(program);
+    return program;
+}
+
+class WebGLNativeRenderer extends Renderer {
+    static get displayName() { return "WebGL Native Renderer"; }
 
     static get description() {
         return "Uses WebGL to render graphics";
@@ -14,78 +65,117 @@ class WebGLRenderer extends Renderer {
 
     static checkSupported() {
         try {
-            var canvas = document.createElement( 'canvas' );
-            return !! (window.WebGLRenderingContext && (canvas.getContext('webgl') || canvas.getContext('experimental-webgl')));
+            const canvas = document.createElement('canvas');
+            return !! (canvas.getContext && (canvas.getContext('webgl') || canvas.getContext('experimental-webgl')));
         } catch(e) {
             return false;
         }
     }
 
-    static get params() {
-        return [
-            Props.component("renderFilter", "Render Filter", RenderFilter)
-        ];
-    }
-
-    constructor() {
-        super();
-        this._renderFilter = new PencilSketchFilter();
-    }
-
-    get renderFilter() {
-        return this._renderFilter;
-    }
-
-    set renderFilter(val) {
-        this._renderFilter = val;
-    }
-
     create(width, height) {
-        this._camera = new Three.OrthographicCamera(-1.5, 1.5, -1.5, 1.5, 0.0001, 1000);
-        this._scene = new Three.Scene();
-        this._renderer = new Three.WebGLRenderer({alpha: true, preserveDrawingBuffer: true });
+        const canvas = document.createElement('canvas');
+        canvas.width = width;
+        canvas.height = height;
 
-        this._camera.position.z = 2;
-        this._renderer.setSize(width, height);
-        this._renderer.setClearColor(0xe0e0e0, 1);
-        return this._renderer.domElement;
+        const gl = canvas.getContext('webgl', { preserveDrawingBuffer: true }) ||
+                   canvas.getContext('experimental-webgl', { preserveDrawingBuffer: true });
+        const buffer = gl.createBuffer();
+
+        const program = createShaderProgram(gl);
+        const positionIndex = gl.getAttribLocation(program, 'position');
+        const colorIndex = gl.getAttribLocation(program, 'color');
+
+        gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+        gl.vertexAttribPointer(positionIndex, 3, gl.FLOAT, false, 28, 0);
+        gl.enableVertexAttribArray(positionIndex);
+
+        gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+        gl.vertexAttribPointer(colorIndex, 4, gl.FLOAT, false, 28, 12);
+        gl.enableVertexAttribArray(colorIndex);
+
+        gl.bindBuffer(gl.ARRAY_BUFFER, null);
+
+        gl.clearColor(0.88, 0.88, 0.88, 1.0);
+        gl.clearDepth(1);
+        gl.enable(gl.DEPTH_TEST);
+        gl.enable(gl.BLEND);
+        gl.blendEquation(gl.FUNC_ADD);
+        gl.blendFunc(gl.ONE, gl.ONE_MINUS_SRC_ALPHA);
+        gl.viewport(0, 0, width, height);
+
+        const projectionMatrix = m4.translate(m4.perspective(degToRad(60), width / height, 0.0001, 100), 0, 0, -2.5);
+        const orthographicMatrix = m4.orthographic(-1.5, 1.5, -1.5, 1.5, -2, 2);
+        const projectionLocation = gl.getUniformLocation(program, 'projection');
+        gl.uniformMatrix4fv(projectionLocation, false, orthographicMatrix);
+
+        this._rotationLocation = gl.getUniformLocation(program, 'rotation');
+
+        this._gl = gl;
+        this._buffer = buffer;
+
+        return canvas;
     }
 
     setRenderData(points) {
-        if (this._geometry) {
-            this._scene.remove(this._cloud);
-            this._geometry.dispose();
+        const gl = this._gl;
+        const buffer = this._buffer;
+
+        const floats = [];
+
+        for (let point of points) {
+            // position
+            for (let i = 0; i < 3; ++i) {
+                floats.push(point[i]);
+            }
+
+            // w
+            //floats.push(1.0);
+
+            // color
+            for (let i = 0; i < 3; ++i) {
+                if (point.length > 3) {
+                    floats.push(point[i]);
+                } else {
+                    floats.push(0.0);
+                }
+            }
+
+            // a
+            if (point.length > 6) {
+                floats.push(point[6]);
+            } else {
+                floats.push(0.1);
+            }
         }
 
-        this._geometry = this._renderFilter.createGeometry(points);
-        let material = this._renderFilter.createMaterial();
-        this._cloud = new Three.PointCloud(this._geometry, material);
-        this._cloud.position.x = 0;
-        this._cloud.position.y = 0;
-        this._cloud.position.z = 0;
+        gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(floats), gl.STATIC_DRAW);
+        gl.bindBuffer(gl.ARRAY_BUFFER, null);
 
-        this._scene.add(this._cloud);
+        this._bufferSize = points.length;
     }
 
     destroy() {
-        if (this._geometry) {
-            this._scene.remove(this._cloud);
-            this._geometry.dispose();
-        }
     }
 
     resize(width, height) {
-        this._renderer.setSize(width, height);
+        this._gl.viewport(0, 0, width, height);
     }
 
     render(rotationX, rotationY) {
-        if (this._cloud) {
-            this._cloud.rotation.x = rotationY;
-            this._cloud.rotation.y = -rotationX;
-            this._renderer.render(this._scene, this._camera);
+        const gl = this._gl;
+
+        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
+
+        const count = this._bufferSize || 0;
+
+        if (count > 0) {
+            const rotationMatrix = m4.xRotate(m4.yRotate(m4.identity(), rotationX), -rotationY);
+            gl.uniformMatrix4fv(this._rotationLocation, false, rotationMatrix);
+            gl.drawArrays(gl.POINTS, 0, count);
         }
     }
 }
 
-Components.register(Renderer, WebGLRenderer, true);
-module.exports = WebGLRenderer;
+Components.register(Renderer, WebGLNativeRenderer, true);
+module.exports = WebGLNativeRenderer;