+++ app/src/js/utils/matrix.js
@@ -1,0 +1,176 @@
+module.exports = {
+    projection: function(width, height, depth) {
+        // Note: This matrix flips the Y axis so 0 is at the top.
+        return [
+            2 / width, 0, 0, 0,
+            0, -2 / height, 0, 0,
+            0, 0, 2 / depth, 0,
+            -1, 1, 0, 1,
+        ];
+    },
+
+    multiply: function(a, b) {
+        const a00 = a[0];
+        const a01 = a[1];
+        const a02 = a[2];
+        const a03 = a[3];
+        const a10 = a[4];
+        const a11 = a[5];
+        const a12 = a[6];
+        const a13 = a[7];
+        const a20 = a[8];
+        const a21 = a[9];
+        const a22 = a[10];
+        const a23 = a[11];
+        const a30 = a[12];
+        const a31 = a[13];
+        const a32 = a[14];
+        const a33 = a[15];
+        const b00 = b[0];
+        const b01 = b[1];
+        const b02 = b[2];
+        const b03 = b[3];
+        const b10 = b[4];
+        const b11 = b[5];
+        const b12 = b[6];
+        const b13 = b[7];
+        const b20 = b[8];
+        const b21 = b[9];
+        const b22 = b[10];
+        const b23 = b[11];
+        const b30 = b[12];
+        const b31 = b[13];
+        const b32 = b[14];
+        const b33 = b[15];
+        return [
+            b00 * a00 + b01 * a10 + b02 * a20 + b03 * a30,
+            b00 * a01 + b01 * a11 + b02 * a21 + b03 * a31,
+            b00 * a02 + b01 * a12 + b02 * a22 + b03 * a32,
+            b00 * a03 + b01 * a13 + b02 * a23 + b03 * a33,
+            b10 * a00 + b11 * a10 + b12 * a20 + b13 * a30,
+            b10 * a01 + b11 * a11 + b12 * a21 + b13 * a31,
+            b10 * a02 + b11 * a12 + b12 * a22 + b13 * a32,
+            b10 * a03 + b11 * a13 + b12 * a23 + b13 * a33,
+            b20 * a00 + b21 * a10 + b22 * a20 + b23 * a30,
+            b20 * a01 + b21 * a11 + b22 * a21 + b23 * a31,
+            b20 * a02 + b21 * a12 + b22 * a22 + b23 * a32,
+            b20 * a03 + b21 * a13 + b22 * a23 + b23 * a33,
+            b30 * a00 + b31 * a10 + b32 * a20 + b33 * a30,
+            b30 * a01 + b31 * a11 + b32 * a21 + b33 * a31,
+            b30 * a02 + b31 * a12 + b32 * a22 + b33 * a32,
+            b30 * a03 + b31 * a13 + b32 * a23 + b33 * a33,
+        ];
+    },
+
+    identity: function() {
+        return [
+            1,  0,  0,  0,
+            0,  1,  0,  0,
+            0,  0,  1,  0,
+            0,  0,  0,  1,
+        ];
+    },
+
+    translation: function(tx, ty, tz) {
+        return [
+            1,  0,  0,  0,
+            0,  1,  0,  0,
+            0,  0,  1,  0,
+            tx, ty, tz, 1,
+        ];
+    },
+
+    xRotation: function(angleInRadians) {
+        const c = Math.cos(angleInRadians);
+        const s = Math.sin(angleInRadians);
+
+        return [
+            1, 0, 0, 0,
+            0, c, s, 0,
+            0, -s, c, 0,
+            0, 0, 0, 1,
+        ];
+    },
+
+    yRotation: function(angleInRadians) {
+        const c = Math.cos(angleInRadians);
+        const s = Math.sin(angleInRadians);
+
+        return [
+            c, 0, -s, 0,
+            0, 1, 0, 0,
+            s, 0, c, 0,
+            0, 0, 0, 1,
+        ];
+    },
+
+    zRotation: function(angleInRadians) {
+        const c = Math.cos(angleInRadians);
+        const s = Math.sin(angleInRadians);
+
+        return [
+            c, s, 0, 0,
+            -s, c, 0, 0,
+            0, 0, 1, 0,
+            0, 0, 0, 1,
+        ];
+    },
+
+    scaling: function(sx, sy, sz) {
+        return [
+            sx, 0,  0,  0,
+            0, sy,  0,  0,
+            0,  0, sz,  0,
+            0,  0,  0,  1,
+        ];
+    },
+
+    translate: function(m, tx, ty, tz) {
+        return this.multiply(m, this.translation(tx, ty, tz));
+    },
+
+    xRotate: function(m, angleInRadians) {
+        return this.multiply(m, this.xRotation(angleInRadians));
+    },
+
+    yRotate: function(m, angleInRadians) {
+        return this.multiply(m, this.yRotation(angleInRadians));
+    },
+
+    zRotate: function(m, angleInRadians) {
+        return this.multiply(m, this.zRotation(angleInRadians));
+    },
+
+    scale: function(m, sx, sy, sz) {
+        return this.multiply(m, this.scaling(sx, sy, sz));
+    },
+
+    perspective: function (fieldOfView, aspect, near, far) {
+        const f = Math.tan(Math.PI * 0.5 - 0.5 * fieldOfView);
+        const rangeInv = 1.0 / (near - far);
+
+        return [
+            f / aspect, 0, 0, 0,
+            0, f, 0, 0,
+            0, 0, (near + far) * rangeInv, -1,
+            0, 0, near * far * rangeInv * 2, 0
+        ];
+    },
+
+    orthographic: function(left, right, bottom, top, near, far) {
+        const w = right - left;
+        const h = top - bottom;
+        const p = far - near;
+
+        const x = ( right + left ) / w;
+        const y = ( top + bottom ) / h;
+        const z = ( far + near ) / p;
+
+        return [
+            2 / w, 0, 0, 0,
+            0, 2 / h, 0, 0,
+            0, 0, -2 / p, 0,
+            -x, -y, -z, 1,
+        ];
+    }
+};